Conflicts:
	indico/web/http_api/api.py

+++ indico/MaKaC/webinterface/rh/base.py
@@ -43,12 +43,13 @@ from MaKaC.webinterface.pages.conferences import WPConferenceModificationClosed
 import MaKaC.webinterface.session as session
 import MaKaC.webinterface.urlHandlers as urlHandlers
 import MaKaC.webinterface.pages.errors as errors
+from MaKaC.webinterface.common.baseNotificator import Notification
 from MaKaC.common.general import *
 
 from MaKaC.accessControl import AccessWrapper
 from MaKaC.common import DBMgr, Config, security
 from MaKaC.errors import MaKaCError, ModificationError, AccessError, KeyAccessError, TimingError, ParentTimingError, EntryTimingError, FormValuesError, NoReportError, NotFoundError, HtmlScriptError, HtmlForbiddenTag, ConferenceClosedError, HostnameResolveError
-from MaKaC.webinterface.mail import GenericMailer, GenericNotification
+from MaKaC.webinterface.mail import GenericMailer
 from xml.sax.saxutils import escape
 
 from MaKaC.common.utils import truncate
@@ -244,7 +245,7 @@ class RH(RequestHandlerBase):
             sm = session.getSessionManager()
             try:
                 self._websession = sm.get_session( self._req )
-            except session.SessionError, e:
+            except session.SessionError:
                 sm.revoke_session_cookie( self._req )
                 self._websession = sm.get_session( self._req )
 
@@ -491,7 +492,8 @@ class RH(RequestHandlerBase):
         profile = Config.getInstance().getProfile()
         proffilename = ""
         res = ""
-        retry = 10
+        MAX_RETRIES = 10
+        retry = MAX_RETRIES
         textLog = []
         self._startTime = datetime.now()
 
@@ -511,18 +513,21 @@ class RH(RequestHandlerBase):
         # notify components that the request has started
         self._notify('requestStarted', self._req)
 
+        forcedConflicts = Config.getInstance().getForceConflicts()
         try:
             while retry>0:
 
-                if retry < 10:
+                if retry < MAX_RETRIES:
                     # notify components that the request is being retried
-                    self._notify('requestRetry', self._req, 10 - retry)
+                    self._notify('requestRetry', self._req, MAX_RETRIES - retry)
 
                 try:
                     Logger.get('requestHandler').info('\t[pid=%s] from host %s' % (os.getpid(), self.getHostIP()))
                     try:
                         # clear the fossile cache at the start of each request
                         fossilize.clearCache()
+                        # delete all queued emails
+                        GenericMailer.flushQueue(False)
 
                         DBMgr.getInstance().sync()
                         # keep a link to the web session in the access wrapper
@@ -562,17 +567,19 @@ class RH(RequestHandlerBase):
 
                         # Save web session, just when needed
                         sm = session.getSessionManager()
-                        sm.maintain_session( self._req, self._websession )
+                        sm.maintain_session(self._req, self._websession)
 
                         # notify components that the request has finished
                         self._notify('requestFinished', self._req)
+                        # Raise a conflict error if enabled. This allows detecting conflict-related issues easily.
+                        if retry > (MAX_RETRIES - forcedConflicts):
+                            raise ConflictError
                         self._endRequestSpecific2RH( True ) # I.e. implemented by Room Booking request handlers
-
                         DBMgr.getInstance().endRequest( True )
                         Logger.get('requestHandler').info('Request %s successful' % (id(self._req)))
                         #request succesfull, now, doing tas that must be done only once
                         try:
-                            self._sendEmails()
+                            GenericMailer.flushQueue(True) # send emails
                             self._deleteTempFiles()
                         except:
                             pass
@@ -582,7 +589,9 @@ class RH(RequestHandlerBase):
                         res = self._processError(e)
                 except (ConflictError, POSKeyError):
                     import traceback
-                    Logger.get('requestHandler').warning('Conflict in Database! (Request %s)\n%s' % (id(self._req), traceback.format_exc()))
+                    # only log conflict if it wasn't forced
+                    if retry <= (MAX_RETRIES - forcedConflicts):
+                        Logger.get('requestHandler').warning('Conflict in Database! (Request %s)\n%s' % (id(self._req), traceback.format_exc()))
                     self._abortSpecific2RH()
                     DBMgr.getInstance().abort()
                     retry -= 1
@@ -711,11 +720,6 @@ class RH(RequestHandlerBase):
             return ""
 
         return res
-
-    def _sendEmails( self ):
-        if hasattr( self, "_emailsToBeSent" ):
-            for email in self._emailsToBeSent:
-                GenericMailer.send(GenericNotification(email))
 
     def _deleteTempFiles( self ):
         if len(self._tempFilesToDelete) > 0:
